src/monitor/requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.26.0
python-dotenv==1.0.1
